backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlmodel==0.0.22
psycopg2-binary==2.9.11
python-jose==3.3.0
better-exceptions==0.3.3
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
bcrypt==4.0.1